Isolating parsing logic from the public interface
